Fixed main.h after chaning variable type in main.c

+++ src/main.h
@@ -4,7 +4,7 @@
 # include <glib.h>
 
 extern GMainContext *game_context;
-extern gulong elapsed_seconds;
+extern guint32 elapsed_seconds;
 
 #endif /* __WMUD_MAIN_H__ */
